Apply suggested changes from PR review
