+++ core/jvm/src/test/scala/scalaz/zio/AbstractRTSSpec.scala
@@ -4,5 +4,5 @@ import org.specs2.Specification
 import scalaz.zio.ExitResult.Cause
 
 trait AbstractRTSSpec extends Specification with RTS {
-  override def defaultHandler: Cause[Nothing] => IO[Nothing, Unit] = _ => IO.unit
+  override def defaultHandler: Cause[Any] => IO[Nothing, Unit] = _ => IO.unit
 }
